// src/enums/error-codes.ts
export enum ErrorCodes {
	GET_USER_FAILED = 'GET_USER_FAILED',
	LIST_USERS_FAILED = 'LIST_USERS_FAILED',
	PUT_USER_FAILED = 'PUT_USER_FAILED',
	UPDATE_USER_FAILED = 'UPDATE_USER_FAILED',
	DELETE_USER_FAILED = 'DELETE_USER_FAILED',
	USER_NOT_FOUND = 'USER_NOT_FOUND',
	DATASOURCE_NOT_INITIALIZED = 'DATASOURCE_NOT_INITIALIZED',
}

export default ErrorCodes;